fix(dto): removed extra dto's & path fix
