+++ src/api/resolvers/test/dto/output/create-sound-addition-output.dto.ts
@@ -1,0 +1,7 @@
+export type CreateSoundAdditionOutput = {
+    "userId": number,
+    "averageCallbackTime": number,
+    "dispersion": number,
+    "allSignals": number,
+    "mistakes": number
+}
