--- src/pages/tests/sound/hard/dto/SoundHardTestAnswer.dto.ts
@@ -1,8 +1,0 @@
-import type {SoundHardTestSumType} from "./SoundHardTestSumType.types.ts";
-
-export type SoundHardTestAnswerDto = {
-    answer: SoundHardTestSumType;
-    questionId: number
-    isCorrect: boolean;
-    elapsedTime: number
-}
