--- src/pages/tests/sound/hard/dto/SoundHardTestQuestion.dto.ts
@@ -1,7 +1,0 @@
-import type {SoundHardTestSumType} from "./SoundHardTestSumType.types.ts";
-
-export type SoundHardTestQuestionDto = {
-    id: number
-    question: string;
-    answer: SoundHardTestSumType;
-}
