--- src/pages/tests/sound/hard/dto/SoundHardTestSumType.types.ts
@@ -1,1 +1,0 @@
-export type SoundHardTestSumType = "ODD" | "EVEN"
